Bookworm REACT
This application has 2 parts. A front end and a backend API. You will have to set up a mongoDB database. The Url is stored in env-sample file in the bookworm-api. This is passed as a parameter in  index.js

The registerServiceWorker.js (in bookworm-react) is used to facilitate performance via cache (on n+1 visit)  and offline.

For more information for help contact by email.  Certain parts of the application has been disabled for security purposes. The application is functional though.